correccion de botones
